K-25133 adjust mock data
