+++ src/mock/html.ts
@@ -1,5 +1,3 @@
-import { anyFunction } from '../lib/util';
-
 type PrototypeType = 'HTML' | 'MouseEvent';
 
 let original: {
@@ -47,71 +45,4 @@ export const unMockProperties = (type: PrototypeType) => {
     original[type].forEach(({ property, value }) => {
       value && Object.defineProperty(prototype, property, value);
     });
-};
-
-export const mockResizeObserver = () => {
-  const observe = jest.fn(),
-    unobserve = jest.fn(),
-    disconnect = jest.fn();
-
-  /**
-   * Trigger for testing purpose only
-   */
-  const triggerResizeEventOnce = (entries?: any[]) => {
-    ResizeObserver._callback(entries ?? ResizeObserver.defaultEntries);
-  };
-
-  class ResizeObserver {
-    static _callback: anyFunction;
-
-    public static defaultWidth = 500;
-    public static defaultHeight = 600;
-    public static defaultInlineAddition = 50;
-
-    public static defaultEntries = [
-      {
-        contentRect: {
-          width: ResizeObserver.defaultWidth,
-          height: ResizeObserver.defaultHeight
-        },
-        borderBoxSize: [
-          {
-            inlineSize: ResizeObserver.defaultWidth + ResizeObserver.defaultInlineAddition,
-            blockSize: ResizeObserver.defaultHeight + ResizeObserver.defaultInlineAddition
-          }
-        ]
-      }
-    ];
-
-    constructor(callback: anyFunction) {
-      ResizeObserver._callback = callback;
-    }
-
-    observe() {
-      observe();
-    }
-
-    unobserve() {
-      unobserve();
-    }
-
-    disconnect() {
-      disconnect();
-    }
-  }
-
-  window.ResizeObserver = ResizeObserver;
-
-  return {
-    observe,
-    unobserve,
-    disconnect,
-    triggerResizeEventOnce,
-    defaults: {
-      width: ResizeObserver.defaultWidth,
-      height: ResizeObserver.defaultHeight,
-      inlineSize: ResizeObserver.defaultWidth + ResizeObserver.defaultInlineAddition,
-      blockSize: ResizeObserver.defaultHeight + ResizeObserver.defaultInlineAddition
-    }
-  };
 };
